+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-##Changelog
+#Changelog
 
-###Gin 0.4 (??)
+###Gin 0.4 (Aug 21, 2014)
+
+- [NEW] Development mode
+- [NEW] Unit tests
+- [NEW] Add Content.Redirect()
+- [FIX] Deferring WriteHeader()
+- [FIX] Improved documentation for model binding
 
 
 ###Gin 0.3 (Jul 18, 2014)